backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
requests==2.31.0
a2wsgi==1.10.0
python-dotenv==1.0.1